Tighten the telemetry suite and drop a redundant field
Four fixes from review:

The telemetry suite was 456 lines against a 400-line limit. The capture
harness is the natural seam, so it moves to `support/tracing_capture.rs`
where neighbouring discovery tests can reach it — the alternative being a
second capture implementation that drifts from the first.

`a_missing_required_candidate_reports_a_required_failure` promised in its
doc comment that the terminal outcome stays `not_found`, then never checked
it. That distinction is the whole point of the test, so it is now asserted.

The metrics assertion totalled `discovery.outcomes` across all labels, which
also asserted that no candidate failed. That is a host-dependent claim —
`/etc/xdg/demo/config.toml` could exist on the machine running the test — so
it now counts the `outcome=not_found` label specifically.

`discovery.candidate` carried `outcome`, `candidate_kind`, and `required`:
three renderings of one boolean. `candidate_kind` is dropped. The original
requirement listed it among example fields rather than mandating it, and
three fields for one bit inflate every event and give a consumer three
things to keep in step.

`default_xdg_group` is split by `cfg` rather than branching in one body:
the non-Unix arm reduces to `Vec::new()`, and the workspace denies
`clippy::missing_const_for_fn`, so a single body would fail Windows Clippy.

Addresses CodeRabbit findings on #411.
