+++ ortho_config/tests/discovery_telemetry.rs
@@ -13,137 +13,15 @@
 //! captured field, so a future field carrying a path fails the suite rather
 //! than shipping.
 
-use anyhow::Context as _;
-use cap_std::{ambient_authority, fs::Dir};
 use ortho_config::{ConfigDiscovery, MapEnv};
+
+#[path = "support/tracing_capture.rs"]
+mod capture_support;
+
+use capture_support::{capture, only, write_fixture};
 use rstest::rstest;
-use std::collections::BTreeMap;
 use std::path::Path;
-use std::sync::{Arc, Mutex, MutexGuard, OnceLock, PoisonError};
-use tracing::field::{Field, Visit};
-use tracing_subscriber::layer::Context;
-use tracing_subscriber::prelude::*;
-use tracing_subscriber::{Layer, Registry};
-
-/// One captured `tracing` event, reduced to its recorded fields.
-#[derive(Debug, Default)]
-struct Captured {
-    fields: BTreeMap<String, String>,
-}
-
-impl Captured {
-    /// Read a field, or the empty string when the event did not record it.
-    fn field(&self, name: &str) -> &str {
-        self.fields.get(name).map_or("", String::as_str)
-    }
-}
-
-#[derive(Default)]
-struct Events(Mutex<Vec<Captured>>);
-
-impl Events {
-    fn lock(&self) -> MutexGuard<'_, Vec<Captured>> {
-        self.0.lock().unwrap_or_else(PoisonError::into_inner)
-    }
-}
-
-struct CaptureLayer(Arc<Events>);
-
-impl<S> Layer<S> for CaptureLayer
-where
-    S: tracing::Subscriber,
-{
-    fn on_event(&self, event: &tracing::Event<'_>, _context: Context<'_, S>) {
-        let mut captured = Captured::default();
-        event.record(&mut FieldVisitor {
-            fields: &mut captured.fields,
-        });
-        self.0.lock().push(captured);
-    }
-}
-
-struct FieldVisitor<'fields> {
-    fields: &'fields mut BTreeMap<String, String>,
-}
-
-impl Visit for FieldVisitor<'_> {
-    fn record_str(&mut self, field: &Field, value: &str) {
-        self.fields
-            .insert(field.name().to_owned(), value.to_owned());
-    }
-
-    fn record_bool(&mut self, field: &Field, value: bool) {
-        self.fields
-            .insert(field.name().to_owned(), value.to_string());
-    }
-
-    fn record_debug(&mut self, field: &Field, value: &dyn std::fmt::Debug) {
-        self.fields
-            .insert(field.name().to_owned(), format!("{value:?}"));
-    }
-}
-
-/// Keep one registered dispatcher alive for the lifetime of the test binary.
-///
-/// `tracing::debug!` consults a *global* maximum level before it ever reaches
-/// the thread-local subscriber, and that maximum is recomputed as dispatchers
-/// register and drop. With tests running concurrently, one thread leaving its
-/// `with_default` scope could drop the global maximum to `OFF` while another
-/// was still inside its own — and the second thread's events would vanish.
-///
-/// This was observed as a genuine intermittent failure, not a theoretical one.
-/// Holding a dispatcher that is never dropped pins the maximum at the registry
-/// default, so no thread can lower it out from under another. It is deliberately
-/// never installed as the default subscriber: each test still captures through
-/// its own thread-local one.
-fn pin_global_max_level() {
-    static PINNED: OnceLock<tracing::Dispatch> = OnceLock::new();
-    let _ = PINNED.get_or_init(|| tracing::Dispatch::new(Registry::default()));
-}
-
-/// Run `f` with a capturing subscriber installed on this thread only.
-///
-/// `with_default` is thread-local, so these tests need no lock and no
-/// `#[serial]` — the same property the injected environment source provides
-/// for the environment itself.
-fn capture<R>(f: impl FnOnce() -> R) -> Vec<Captured> {
-    pin_global_max_level();
-    let events = Arc::new(Events::default());
-    let subscriber = Registry::default().with(CaptureLayer(Arc::clone(&events)));
-    tracing::subscriber::with_default(subscriber, f);
-    std::mem::take(&mut *events.lock())
-}
-
-/// Every captured event carrying the given `event` name.
-fn named<'a>(events: &'a [Captured], name: &str) -> Vec<&'a Captured> {
-    events
-        .iter()
-        .filter(|event| event.field("event") == name)
-        .collect()
-}
-
-/// Exactly one event with `name` must have been emitted; return it.
-fn only<'a>(events: &'a [Captured], name: &str) -> &'a Captured {
-    let matching = named(events, name);
-    match matching.as_slice() {
-        [event] => event,
-        other => panic!("expected exactly one `{name}` event, got {other:?}"),
-    }
-}
-
-/// Write a minimal loadable configuration file into `dir`, returning its path.
-///
-/// The write goes through a `cap_std::fs::Dir` handle rather than `std::fs`,
-/// which the repository's lint suite requires: a capability handle names the
-/// directory it may touch, so a fixture cannot accidentally write relative to
-/// the process's working directory.
-fn write_fixture(dir: &std::path::Path, name: &str) -> anyhow::Result<std::path::PathBuf> {
-    let cap =
-        Dir::open_ambient_dir(dir, ambient_authority()).context("open the temporary directory")?;
-    cap.write(name, b"value = 1\n")
-        .context("write the fixture")?;
-    Ok(dir.join(name))
-}
+use std::sync::Arc;
 
 fn discovery_with(env: MapEnv) -> ConfigDiscovery {
     ConfigDiscovery::builder("demo")
@@ -318,9 +196,15 @@ fn a_missing_required_candidate_reports_a_required_failure() {
 
     let events = capture(|| discovery.load_first_partitioned());
 
+    // The doc comment above promises the terminal outcome stays `not_found`;
+    // pin it, since that distinction is the whole point of the test.
+    assert_eq!(
+        only(&events, "discovery.load").field("outcome"),
+        "not_found"
+    );
+
     let candidate = only(&events, "discovery.candidate");
     assert_eq!(candidate.field("outcome"), "required_failure");
-    assert_eq!(candidate.field("candidate_kind"), "required");
     assert_eq!(candidate.field("required"), "true");
 }
 
@@ -414,21 +298,41 @@ mod metrics_facade {
 
     use super::{MapEnv, discovery_with};
     use metrics_util::debugging::{DebugValue, DebuggingRecorder, Snapshot};
-    use std::collections::BTreeMap;
 
-    /// Total each counter across its label sets.
+    /// Count one counter, restricted to a single label.
     ///
-    /// Summing over labels is deliberate: the assertion is about how many
-    /// discovery operations were counted, not about the label vocabulary,
-    /// which the tracing tests above already pin.
-    fn counter_totals(snapshot: Snapshot) -> BTreeMap<String, u64> {
-        let mut totals = BTreeMap::new();
-        for (key, _, _, value) in snapshot.into_vec() {
-            if let DebugValue::Counter(count) = value {
-                *totals.entry(key.key().name().to_owned()).or_insert(0_u64) += count;
-            }
-        }
-        totals
+    /// Scoped by label rather than totalled: `count_outcome` fires from both
+    /// `candidate_failure` and `load_outcome`, so a bare total would also be
+    /// asserting that no candidate failed. That is a host-dependent claim —
+    /// `/etc/xdg/demo/config.toml` is a real path that could exist on the
+    /// machine running the test — and it is not what this test is about.
+    fn counter_for(
+        entries: &[(metrics_util::CompositeKey, u64)],
+        name: &str,
+        label: (&str, &str),
+    ) -> u64 {
+        entries
+            .iter()
+            .filter(|(key, _)| {
+                key.key().name() == name
+                    && key
+                        .key()
+                        .labels()
+                        .any(|found| found.key() == label.0 && found.value() == label.1)
+            })
+            .map(|(_, count)| *count)
+            .sum()
+    }
+
+    fn counters(snapshot: Snapshot) -> Vec<(metrics_util::CompositeKey, u64)> {
+        snapshot
+            .into_vec()
+            .into_iter()
+            .filter_map(|(key, _, _, value)| match value {
+                DebugValue::Counter(count) => Some((key, count)),
+                _ => None,
+            })
+            .collect()
     }
 
     #[test]
@@ -441,16 +345,24 @@ mod metrics_facade {
             drop(discovery.load_first());
         });
 
-        let totals = counter_totals(snapshotter.snapshot());
+        let entries = counters(snapshotter.snapshot());
         assert_eq!(
-            totals.get("ortho_config.discovery.attempts").copied(),
-            Some(1),
-            "one discovery attempt should be counted, got {totals:?}"
+            counter_for(
+                &entries,
+                "ortho_config.discovery.attempts",
+                ("operation", "discover_first")
+            ),
+            1,
+            "one discovery attempt should be counted, got {entries:?}"
         );
         assert_eq!(
-            totals.get("ortho_config.discovery.outcomes").copied(),
-            Some(1),
-            "one terminal outcome should be counted, got {totals:?}"
+            counter_for(
+                &entries,
+                "ortho_config.discovery.outcomes",
+                ("outcome", "not_found")
+            ),
+            1,
+            "one terminal not_found outcome should be counted, got {entries:?}"
         );
     }
 }
